:recycle: Compose 사분면 모듈 분리
