+++ Unit1/Pathway3/TaskCompleted/app/src/main/java/com/example/taskcompleted/MainActivity.kt
@@ -34,72 +34,13 @@ class MainActivity : ComponentActivity() {
             TaskCompletedTheme() {
 //                ComposeHelper()
 //                TaskCompletedScreen()
-                ComposeQuadrant()
+
             }
         }
     }
 }
 
-@Composable
-fun ComposeQuadrant() {
-    Row(Modifier.fillMaxHeight()) {
-        Column(modifier = Modifier.weight(1f)) {
-            ComposeCard(
-                "Text composable",
-                "Displays text and follows Material Design guidelines.",
-                Color.Green,
-                modifier = Modifier.weight(1f)
-            )
-            ComposeCard(
-                "Row composable",
-                "A layout composable that places its children in a horizontal sequence.",
-                Color.Cyan,
-                modifier = Modifier.weight(1f)
-            )
-        }
-        Column(modifier = Modifier.weight(1f)) {
-            ComposeCard(
-                "Image composable",
-                "Creates a composable that lays out and draws a given Painter class object.",
-                Color.Yellow,
-                modifier = Modifier.weight(1f)
-            )
-            ComposeCard(
-                "Column composable",
-                "A layout composable that places its children in a vertical sequence.",
-                Color.LightGray,
-                modifier = Modifier.weight(1f)
-            )
-        }
-    }
-}
 
-@Composable
-fun ComposeCard(
-    title: String,
-    content: String,
-    backgroundColor: Color,
-    modifier: Modifier
-) {
-    Column(
-        modifier
-            .fillMaxSize()
-            .background(backgroundColor)
-            .padding(16.dp),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(
-            text = title,
-            fontWeight = FontWeight.Bold,
-            modifier = Modifier.padding(bottom = 16.dp)
-        )
-        Text(
-            text = content,
-            textAlign = TextAlign.Justify
-        )
-    }
-}
 
 @Composable
 fun TaskCompletedScreen() {
